+++ packages/dust_flutter/lib/src/router/routing_core.dart
@@ -9,6 +9,7 @@ part 'routing_delegate.dart';
 part 'routing_delegate_diagnostics.dart';
 part 'routing_delegate_stack.dart';
 part 'routing_guards.dart';
+part 'routing_helpers.dart';
 part 'routing_page.dart';
 part 'routing_parser.dart';
 part 'routing_types.dart';